[RUNTIME] Adds Basic Protobuf Packaging
Adds actual packaging logic to the package method in Ansible.py. It currently parses through
state and finds the studentCodeStatus as well as a limit switch that has default values for
testing. These values have been hardcoded in but should be mutable when we do full integration
testing.

This branch will have compiled protos for testing purposes but the final commit/pull request
will have this removed and test if devops can compile it themselves.

Runtime.proto has also been changed to add an ESTOP enum as a potential state and a uid field
to potentially send UID.

This code has only been tested with fake_dawn