[Runtime] Adds Protobuf Integration
Adds packaging and unpackaging functionality for the UDP send and receive
processes for google protobufs.

Mirrors packaging and unpackaging code given by the protobuf python guide.

For packaging, creates protobuf objects, parses through the state, and adds
data to the correct fields of the objects. It then serializes the data to bytes
and gets sent by the socket.

Due to the current fields in the state dictionary, we only serialize a single limit
switch but it can be expanded to all other types of sensors in the state.

For unpackaging, it reads and creates a protobuf object from serialized data and
then parses it and adds it to fields in a dictionary which is then added to the state.

fake_dawn also mirrors this but with the ansible proto.